// src/core/enums/PokemonType.ts
export enum PokemonType {
  GRASS = 'Grass',
  POISON = 'Poison',
  FIRE = 'Fire',
  WATER = 'Water',
  ELECTRIC = 'Electric',
  FLYING = 'Flying',
  NORMAL = 'Normal',
  FAIRY = 'Fairy',
  PSYCHIC = 'Psychic',
  FIGHTING = 'Fighting',
  GHOST = 'Ghost',
  ROCK = 'Rock',
  GROUND = 'Ground',
  DRAGON = 'Dragon',
  STEEL = 'Steel',
}